jitsi/tests/requirements.txt
# Testing framework
pytest==8.4.2
pytest-cov==4.1.0
pytest-timeout==2.2.0
pytest-html==4.1.1
pytest-xdist==3.5.0

# Selenium y drivers
selenium==4.15.2
webdriver-manager==4.0.1

# Docker client
docker==7.0.0

# Utilidades
python-dotenv==1.0.0
requests==2.31.0
